+++ R/read_chroms.R
@@ -2,15 +2,17 @@
 #'
 #' Reads chromatograms from specified folders or vector of paths using file parsers
 #' from [Aston](https://github.com/bovee/aston), [Entab](https://github.com/bovee/entab),
-#' and [ThermoRawFileParser](https://github.com/compomics/ThermoRawFileParser).
+#' [ThermoRawFileParser](https://github.com/compomics/ThermoRawFileParser),
+#' [OpenChrom](https://lablicate.com/platform/openchrom), and
+#' [rainbow](https://rainbow-api.readthedocs.io/).
 #'
 #' Provides a general interface to chromConverter parsers. Currently recognizes
 #' 'Agilent ChemStation' (.uv), 'MassHunter' (.dad) files, 'Thermo RAW',
-#' 'Waters ARW' (.arw), 'Chromeleon ASCII' (.txt), 'Shimadzu ASCII' (.txt). Also,
-#' wraps Openchrom parsers, which include many additional formats. To use 'Entab',
-#' 'ThermoRawFileParser', or 'Openchrom' parsers, they must be manually installed.
-#' Please see the instructions in the [README](https://ethanbass.github.io/chromConverter/)
-#' for further details.
+#' 'Waters ARW' (.arw), 'Waters RAW' (.raw), 'Chromeleon ASCII' (.txt),
+#' 'Shimadzu ASCII' (.txt). Also, wraps Openchrom parsers, which include many
+#' additional formats. To use 'Entab', 'ThermoRawFileParser', or 'Openchrom'
+#' parsers, they must be manually installed. Please see the instructions in the
+#' [README](https://ethanbass.github.io/chromConverter/) for further details.
 #'
 #' @name read_chroms
 #' @param paths paths to files or folders containing files
@@ -18,14 +20,18 @@
 #' the function with a folder or vector of folders containing the files.
 #' Otherwise, set to\code{FALSE}.
 #' @param format_in Format of files to be imported/converted. The current options
-#' are: \code{chemstation_uv}, \code{chemstation_csv}, \code{masshunter_dad}, \code{shimadzu_fid},
-#' \code{shimadzu_dad}, \code{chromeleon_uv}, \code{thermoraw}, \code{mzml},
-#' \code{waters_arw}, \code{msd}, \code{csd}, \code{wsd}, or \code{other}.
+#' are: \code{chemstation_uv}, \code{chemstation}, \code{chemstation_csv},
+#' \code{masshunter}, \code{masshunter_dad}, \code{shimadzu_fid}, \code{shimadzu_dad},
+#' \code{chromeleon_uv}, \code{agilent_d}, \code{thermoraw}, \code{mzml},
+#' \code{waters_arw}, \code{waters_raw}, \code{msd}, \code{csd}, \code{wsd},
+#' or \code{other}.
 #' @param pattern pattern (e.g. a file extension). Defaults to NULL, in which
 #' case file extension will be deduced from \code{format_in}.
 #' @param parser What parser to use. Current option are \code{chromconverter},
-#' \code{aston}, \code{entab}, \code{thermoraw}, or \code{openchrom}.
+#' \code{aston}, \code{entab}, \code{thermoraw}, \code{openchrom}, or \code{rainbow}.
 #' @param format_out R object format (i.e. data.frame or matrix).
+#' @param data_format Whether to output data in wide or long format. Either
+#' \code{wide} or \code{long}.
 #' @param export Logical. If TRUE, will export files as csvs.
 #' @param path_out Path for exporting files. If path not specified, files will
 #' export to current working directory.
@@ -52,27 +58,37 @@
 #' @export read_chroms
 
 read_chroms <- function(paths, find_files,
-                        format_in=c("chemstation_uv", "chemstation_csv", "masshunter_dad",
+                        format_in=c("agilent_d", "chemstation", "chemstation_uv",
+                                    "chemstation_csv", "masshunter_dad",
                                     "shimadzu_fid", "shimadzu_dad", "chromeleon_uv",
-                                   "thermoraw", "mzml", "waters_arw", "msd",
-                                   "csd", "wsd", "other"),
+                                   "thermoraw", "mzml", "waters_arw", "waters_raw",
+                                   "msd", "csd", "wsd", "other"),
                         pattern = NULL,
                         parser = c("", "chromconverter", "aston", "entab",
-                                   "thermoraw", "openchrom"),
+                                   "thermoraw", "openchrom", "rainbow"),
                         format_out = c("matrix", "data.frame"),
+                        data_format = c("wide","long"),
                         export = FALSE, path_out = NULL,
                         export_format = c("csv", "cdf", "mzml", "animl"),
                         read_metadata = TRUE, dat = NULL){
-  format_in <- match.arg(format_in, c("chemstation_uv", "chemstation_csv",
-                                      "masshunter_dad", "shimadzu_fid",
-                                      "shimadzu_dad", "chromeleon_uv",
+  format_in <- match.arg(format_in, c("agilent_d", "chemstation", "chemstation_uv",
+                                      "chemstation_csv", "masshunter_dad",
+                                      "shimadzu_fid", "shimadzu_dad", "chromeleon_uv",
                                       "thermoraw", "mzml", "waters_arw",
-                                        "msd", "csd", "wsd", "other"))
+                                      "waters_raw", "msd", "csd", "wsd", "other"))
+  data_format <- match.arg(data_format, c("wide","long"))
   format_out <- match.arg(format_out, c("matrix", "data.frame"))
   parser <- match.arg(parser, c("", "chromconverter", "aston","entab",
-                                          "thermoraw", "openchrom"))
+                                          "thermoraw", "openchrom", "rainbow"))
   if (missing(find_files)){
-    ft <- all(file_test("-f",paths))
+    if (!(format_in %in% c("agilent_d", "waters_raw"))){
+      ft <- all(file_test("-f", paths))
+    } else{
+      ext <- switch(format_in,
+                        agilent_d = "\\.d",
+                        waters_raw = "\\.raw")
+      ft <- all(grepl(ext, paths, ignore.case = TRUE))
+    }
     find_files <- !ft
   }
   if (parser == ""){
@@ -106,19 +122,34 @@ read_chroms <- function(paths, find_files,
     dat<-list()}
 
   # choose converter
-  if (format_in == "masshunter_dad"){
+  entab_parser <- partial(call_entab, format_in = format_in,
+                          format_out = format_out,
+                          data_format = data_format,
+                          read_metadata = read_metadata)
+
+  rainbow_parser <- partial(call_rainbow, format_in = format_in,
+          format_out = format_out, data_format = data_format,
+          read_metadata = read_metadata)
+
+  if (format_in == "agilent_d"){
+    pattern <- ifelse(is.null(pattern), ".D", pattern)
+    converter <- rainbow_parser
+  } else if (format_in == "masshunter_dad"){
     pattern <- ifelse(is.null(pattern), ".sp", pattern)
     converter <- switch(parser,
-                        "aston" = sp_converter,
-                        "entab" = partial(call_entab, read_metadata = read_metadata, format_out = format_out,
-                                          format_in = format_in))
+                        "aston" = partial(sp_converter, format_out = format_out,
+                                          data_format = data_format,
+                                          read_metadata = read_metadata),
+                        "entab" = entab_parser)
   } else if (format_in == "chemstation_uv"){
     pattern <- ifelse(is.null(pattern), ".uv", pattern)
     converter <- switch(parser,
-                        "aston" = partial(uv_converter, read_metadata = read_metadata, format_out = format_out),
-                        "entab" = partial(call_entab, read_metadata = read_metadata, format_out = format_out,
-                                          format_in = format_in))
-  } else if(format_in == "chromeleon_uv"){
+                        "aston" = partial(uv_converter, format_out = format_out,
+                                          data_format = data_format,
+                                          read_metadata = read_metadata),
+                        "entab" = entab_parser,
+                        "rainbow" = rainbow_parser)
+  } else if (format_in == "chromeleon_uv"){
     pattern <- ifelse(is.null(pattern), ".txt", pattern)
     converter <- partial(read_chromeleon, read_metadata = read_metadata, format_out = format_out)
   } else if (format_in == "shimadzu_fid"){
@@ -135,14 +166,16 @@ read_chroms <- function(paths, find_files,
                     "thermoraw" = partial(read_thermoraw, path_out = path_out,
                                             read_metadata = read_metadata,
                                             format_out = format_out),
-                     "entab" = partial(call_entab, read_metadata = read_metadata,
-                                        format_out = format_out, format_in = format_in))
+                     "entab" = entab_parser)
   } else if (format_in == "mzml"){
     pattern <- ifelse(is.null(pattern), ".mzML", pattern)
     converter <- partial(read_mzml, format_out = format_out)
   } else if (format_in == "waters_arw"){
     pattern <- ifelse(is.null(pattern), ".arw", pattern)
     converter <- partial(read_waters_arw, format_out = format_out)
+  } else if (format_in == "waters_raw"){
+      pattern <- ifelse(is.null(pattern), ".raw", pattern)
+      converter <- rainbow_parser
   } else if (format_in == "chemstation_csv"){
     pattern <- ifelse(is.null(pattern), ".csv|.CSV", pattern)
     converter <- partial(read_chemstation_csv, format_out = format_out)
@@ -156,7 +189,9 @@ read_chroms <- function(paths, find_files,
                          return_paths = return_paths)
   } else{
     converter <- switch(parser,
-                        "aston" = trace_converter,
+                        "aston" = partial(trace_converter, format_out = format_out,
+                                          data_format = data_format,
+                                          read_metadata = read_metadata),
                         "entab" = partial(call_entab,
                                           read_metadata = read_metadata,
                                           format_out = format_out)