+++ NEWS.md
@@ -1,4 +1,4 @@
-## chromConverter 0.2.3
+## chromConverter 0.3.0
 
 * Fixed bug causing "Chromeleon" metadata parser to fail.
 * Fixed bug in "Chemstation" metadata parser.
@@ -6,6 +6,8 @@
 * Improved error-handling for parsing metadata so small problems no longer error out the whole program.
 * Changed `format_data` argument to `data_format` to select wide or long format.
 * Added support for parsing `mzML` files with `RaMS`.
+* Added support for parsing "Agilent" (`.D`) and "Waters" (`.raw`) files with [rainbow](https://rainbow-api.readthedocs.io/).
+* Make `data_format` option available consistently for choosing `wide` or `long` format.
 
 ## chromConverter 0.2.2